task3 vers 1

+++ src/main/java/TaskThree/DateIn.java
@@ -5,5 +5,5 @@ public interface DateIn {
     boolean isValidDate(int year,int month,int day);
     int getDayOfWeek(int year,int month,int day);
     String toString(int year,int month,int day);
-    int CountDays(int year,int month,int day);
+    long CountDays(int year,int month,int day);
 }
